backend/requirements.txt
Django==4.2.3
djangorestframework==3.14.0
django-cors-headers==4.2.0
python-dotenv==1.0.0

requests>=2.31.0
gitpython>=3.1.40
chardet>=5.2.0